requirements.txt
anyio==3.6.2
certifi @ file:///C:/b/abs_85o_6fm0se/croot/certifi_1671487778835/work/certifi
click==8.1.3
colorama==0.4.6
dnspython==2.3.0
fastapi==0.89.1
h11==0.14.0
idna==3.4
mypy-extensions==0.4.3
passlib==1.7.4
pathspec==0.11.0
platformdirs==2.6.2
pydantic==1.10.4
PyJWT==2.6.0
pymongo==4.3.3
python-decouple==3.7
sniffio==1.3.0
starlette==0.22.0
tomli==2.0.1
typing_extensions==4.4.0
uvicorn==0.20.0
wincertstore==0.2
